+++ src/messages/es/admin.php
@@ -349,4 +349,15 @@ return [
     'js_pagination_page' => 'Página',
     'dashboard_log_message_delete' => 'Registro borrado en <strong>{container}</strong>.',
     'model_tag_relations_count' => 'Usado',
+
+// 1.2.3
+
+    'active_button_duplicate_label' => 'Duplicate',
+    'active_button_duplicate_success' => 'A copy has been created.',
+    'active_button_duplicate_error' => 'Error while creating the copy: {message}',
+
+// 1.3
+
+'model_tag_translation' => 'Translation',
+'model_tag_translation_hint' => 'An optional translation for the tag in the given language. This allows to assign and output the same tag on multi-language websites.'
 ];